moved package for dao/repository to make it more logical; added assertions to test: db insert and update
