+++ src/main/java/ca/jent/dddjdbccomplex/aggregates/subscriptionoffer/dao/SubscriptionOfferRepository.java
@@ -1,9 +1,9 @@
-package ca.jent.dddjdbccomplex.dao.subscriptionoffer;
+package ca.jent.dddjdbccomplex.aggregates.subscriptionoffer.dao;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import static ca.jent.dddjdbccomplex.dao.subscriptionoffer.SubscriptionOfferDao.*;
+import static ca.jent.dddjdbccomplex.aggregates.subscriptionoffer.dao.SubscriptionOfferDao.*;
 
 @Repository
 public interface SubscriptionOfferRepository extends CrudRepository<SubscriptionOfferEntity, Long> {
